requirements-production.txt
# Core Flask and Web Framework
flask==3.0.3
werkzeug==3.0.1
jinja2==3.1.2
markupsafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Database and ORM
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
alembic==1.13.1
flask-migrate==4.0.5

# PostgreSQL support for production
psycopg2-binary==2.9.9

# Authentication and Forms
flask-login==0.6.3
flask-wtf==1.2.1
wtforms==3.1.1

# Email functionality
flask-mail==0.9.1
blinker==1.7.0

# Image processing for photo uploads
pillow==10.2.0

# Environment configuration
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Development and utilities
python-dateutil==2.8.2